requirements.txt
fastapi
uvicorn
pillow
ultralytics
torch
